adapted jest test to deploy in the not jenkins case
