+++ tests/create.test.js
@@ -1,0 +1,16 @@
+const childProcess = require('child_process')
+const _ = require('lodash')
+
+const Teleport = require('../lib').default
+const { TEST_APP_NAME, TEST_TEMPLATES, TEST_APP_DIR } = require('./test.config.js')
+
+// equivalent as 'tpt -c --name ${TEST_APP_NAME} --templates ${TEST_TEMPLATES}'
+test('create task', () => {
+  childProcess.execSync(`rm -rf ${TEST_APP_NAME}`)
+  const testTeleport = new Teleport({
+    create: true,
+    name: TEST_APP_NAME,
+    templates: TEST_TEMPLATES
+  })
+  testTeleport.launch()
+})
