+++ tests/jest_test
@@ -1,0 +1,1 @@
+export JEST_TESTER=jenkins && npm run test
